requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.1.0
sqlparse==0.5.1
tzdata==2024.2
